+++ server/portfolio-os/middleware/authConfig.js
@@ -14,7 +14,8 @@ const GoogleStrategy = require( 'passport-google-oauth20' ).Strategy;
 const userStore = require( '../dataAccess/userStore' );
 const db = require( '../../db' );
 const verifyJWT = require( './verifyJWT' );
-const { sanitizeHTML } = require('../../../common/commonUtils');
+const { sanitizeHTML } = require( '../../../common/commonUtils' );
+const setResetCookie = require( '../middleware/setResetCookie' );
 const LoginType = require( '../../../common/enums/loginType' );
 
 module.exports = ( app ) => {
@@ -75,6 +76,8 @@ module.exports = ( app ) => {
 
       // This catch block catches any possible exceptions that may occure from above.
     } catch ( e ) {
+      setResetCookie( req );
+
       // return done( e, null );
       return done( null, null, { message: 'Unknown Error.' } );
     }
@@ -122,6 +125,8 @@ module.exports = ( app ) => {
       return done( null, user );
 
     } catch ( e ) {
+      setResetCookie( req );
+
       // return done( e, null );
       return done( null, null, { message: 'Unknown Error.' } );
     }
@@ -152,6 +157,7 @@ const __linkAccountToUserIfTrue = ( req, profileId ) => {
       return resolve( false );
 
     } catch ( e ) {
+      setResetCookie( req );
       return reject( e );
     }
   } );