+++ server/portfolio-os/middleware/setResetCookie.js
@@ -1,0 +1,11 @@
+﻿module.exports = ( req ) => {
+  req.cookie(
+    'reset',
+    true,
+    {
+      signed: true,
+      audience: process.env.JWT_AUDIENCE,
+      issuer: process.env.JWT_ISSUER
+    }
+  );
+};